+++ src/plugins/nonebot_bison/platform/bilibili.py
@@ -2,11 +2,10 @@ import json
 import re
 from typing import Any, Optional
 
-import httpx
-
 from ..post import Post
 from ..types import Category, RawPost, Tag, Target
-from .platform import CategoryNotSupport, NewMessage, Platform
+from ..utils import http_client
+from .platform import CategoryNotSupport, NewMessage
 
 
 class Bilibili(NewMessage):
@@ -30,7 +29,7 @@ class Bilibili(NewMessage):
     parse_target_promot = "请输入用户主页的链接"
 
     async def get_target_name(self, target: Target) -> Optional[str]:
-        async with httpx.AsyncClient() as client:
+        async with http_client() as client:
             res = await client.get(
                 "https://api.bilibili.com/x/space/acc/info", params={"mid": target}
             )
@@ -47,10 +46,10 @@ class Bilibili(NewMessage):
         ):
             return Target(match.group(1))
         else:
-            raise Platform.ParseTargetException()
+            raise self.ParseTargetException()
 
     async def get_sub_list(self, target: Target) -> list[RawPost]:
-        async with httpx.AsyncClient() as client:
+        async with http_client() as client:
             params = {"host_uid": target, "offset": 0, "need_top": 0}
             res = await client.get(
                 "https://api.vc.bilibili.com/dynamic_svr/v1/dynamic_svr/space_history",