+++ src/plugins/nonebot_bison/platform/ncm_artist.py
@@ -1,10 +1,9 @@
 import re
 from typing import Any, Optional
 
-import httpx
-
 from ..post import Post
 from ..types import RawPost, Target
+from ..utils import http_client
 from .platform import NewMessage
 
 
@@ -22,7 +21,7 @@ class NcmArtist(NewMessage):
     parse_target_promot = "请输入歌手主页（包含数字ID）的链接"
 
     async def get_target_name(self, target: Target) -> Optional[str]:
-        async with httpx.AsyncClient() as client:
+        async with http_client() as client:
             res = await client.get(
                 "https://music.163.com/api/artist/albums/{}".format(target),
                 headers={"Referer": "https://music.163.com/"},
@@ -43,7 +42,7 @@ class NcmArtist(NewMessage):
             raise self.ParseTargetException()
 
     async def get_sub_list(self, target: Target) -> list[RawPost]:
-        async with httpx.AsyncClient() as client:
+        async with http_client() as client:
             res = await client.get(
                 "https://music.163.com/api/artist/albums/{}".format(target),
                 headers={"Referer": "https://music.163.com/"},