+++ src/plugins/nonebot_bison/platform/weibo.py
@@ -3,13 +3,13 @@ import re
 from datetime import datetime
 from typing import Any, Optional
 
-import httpx
 from bs4 import BeautifulSoup as bs
 from nonebot.log import logger
 
 from ..post import Post
 from ..types import *
-from .platform import NewMessage, Platform
+from ..utils import http_client
+from .platform import NewMessage
 
 
 class Weibo(NewMessage):
@@ -31,7 +31,7 @@ class Weibo(NewMessage):
     parse_target_promot = "请输入用户主页（包含数字UID）的链接"
 
     async def get_target_name(self, target: Target) -> Optional[str]:
-        async with httpx.AsyncClient() as client:
+        async with http_client() as client:
             param = {"containerid": "100505" + target}
             res = await client.get(
                 "https://m.weibo.cn/api/container/getIndex", params=param
@@ -49,10 +49,10 @@ class Weibo(NewMessage):
             # 都2202年了应该不会有http了吧，不过还是防一手
             return Target(match.group(1))
         else:
-            raise Platform.ParseTargetException()
+            raise self.ParseTargetException()
 
     async def get_sub_list(self, target: Target) -> list[RawPost]:
-        async with httpx.AsyncClient() as client:
+        async with http_client() as client:
             params = {"containerid": "107603" + target}
             res = await client.get(
                 "https://m.weibo.cn/api/container/getIndex?", params=params, timeout=4.0
@@ -138,7 +138,7 @@ class Weibo(NewMessage):
             retweeted = True
         pic_num = info["retweeted_status"]["pic_num"] if retweeted else info["pic_num"]
         if info["isLongText"] or pic_num > 9:
-            async with httpx.AsyncClient() as client:
+            async with http_client() as client:
                 res = await client.get(
                     "https://m.weibo.cn/detail/{}".format(info["mid"]), headers=header
                 )